Phase 1: Split core and shared DTOs into schema and class files
