+++ src/app/types/app.types.ts
@@ -1,4 +1,4 @@
-import type { APP_METADATA_DTO } from "@app/dto/app-metadata/app-metadata.dto";
+import type { APP_METADATA_DTO } from "@app/dto/app-metadata/app-metadata.schema";
 
 import type { z } from "zod";
